mv and separate
