--- Assets/UniGLTF/Runtime/UniGLTF/IO/TextureConverter.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 5d2d27884d6c4780b08c745e5cc5c5b5
-timeCreated: 1537261040
