+++ Assets/UniGLTF/Runtime/UniGLTF/IO/TextureConverter.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 79721770f23ba6748abc1720cd99ad74
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
