+++ Assets/UniGLTF/Runtime/UniGLTF/IO/TextureConverter/ITextureConverter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 55c6531d46a98a845b19cbe1f938eab1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
