+++ Assets/UniGLTF/Runtime/UniGLTF/IO/TextureConverter/NormalConverter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5c6170d9d89ac7f43b6860805e5e6214
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
