+++ Assets/UniGLTF/Runtime/UniGLTF/IO/TextureConverter/TextureConverter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0342907cb8901f44b8792016ac152fcd
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
